Addition to saveOrder

+++ source/goodway/goodway-domain/src/main/java/mksgroup/goodway/repository/ProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import mksgroup.goodway.entity.Product;
 
 @Repository
-public interface ProductRepository extends CrudRepository<Product, Long> {
+public interface ProductRepository extends CrudRepository<Product, Integer> {
     
 }
